Configure filesystem paths from the command line
Currently, all paths used by the server are hard-coded to paths in the
directory where the server source code is located. On Unix systems, it's
preferable to store state, configuration, and log files in different
directories. Therefore, this commit adds the following new command-line
arguments to configure these paths:

- `--state-dir`: configures the directory where the database and
  uploaded image files are stored.
- `--config-path`: configures the path to the config file.
- `--log-path`: configures the path to the directory where log files
  will be written.
- `--lock-path`: configures the path to the directory where the server
  will store its PID lock file.

If these command-line arguments are not provided, the server behaves
identically to the way it does today.