Fix `force` parameter for set image endpoint not working
* Tornado's `RequestHandler.get_argument()` returns either a string or
  `None`.
* The type hint for the default value is `str`, but `False` is passed.

Thus, the original comparison against `True` is not appropriate as, at
least when a value for the query parameter is given, the value will
never be of type `bool`.

Instead, compare against a string and use a string as default value to
fix this.

I opted to accept `true` as the only string value that sets the force
flag, but this could be extended to ignore case or allow other values.